do-over of checkpoint 3
