+++ payable.rb
@@ -1,0 +1,12 @@
+module Payable
+
+	def salary=(new_salary)
+	    puts "This better be good!"
+	    @salary = new_salary
+	end
+
+	def receive_raise(raise)
+	  	@salary += raise
+	end
+
+end
