+++ spec/senior_teacher_spec.rb
@@ -1,4 +1,7 @@
 require_relative '../senior_teacher'
+require_relative '../person'
+require_relative '../payable'
+
 
 describe SeniorTeacher do
   let(:jan) { SeniorTeacher.new({name: "Jan", age: 35}) }
